Try  move install to `setup.js`
