+++ jest.config.js
@@ -1,21 +1,8 @@
 "use strict";
 
-const path = require("path");
-const installPrettier = require("./scripts/install-prettier");
-
-const PROJECT_ROOT = __dirname;
 const isProduction = process.env.NODE_ENV === "production";
-const ENABLE_CODE_COVERAGE = Boolean(process.env.ENABLE_CODE_COVERAGE);
 const TEST_STANDALONE = Boolean(process.env.TEST_STANDALONE);
-const INSTALL_PACKAGE = Boolean(process.env.INSTALL_PACKAGE);
-
-let PRETTIER_DIR = isProduction
-  ? path.join(PROJECT_ROOT, "dist")
-  : PROJECT_ROOT;
-if (INSTALL_PACKAGE || (isProduction && !TEST_STANDALONE)) {
-  PRETTIER_DIR = installPrettier(PRETTIER_DIR);
-}
-process.env.PRETTIER_DIR = PRETTIER_DIR;
+const ENABLE_CODE_COVERAGE = Boolean(process.env.ENABLE_CODE_COVERAGE);
 
 const testPathIgnorePatterns = [];
 let transform;
